gh 0.0.0 Add index, Readme, License and fix the whole project for fun
