+++ .gitignore
@@ -1,0 +1,3 @@
+node_modules
+package-lock.json
+pnpm-lock.yaml
