+++ LICENSE.md
@@ -1,0 +1,14 @@
+Pylar AI Creative ML Free License
+Copyright (C) 2023 Laiba Ahmed and Miguel Gargallo
+
+Permission
+Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+
+Conditions
+The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+Any work using, copying, modifying, merging, publishing, distributing, sublicensing, and/or selling copies of the Software must explicitly acknowledge "Miguel Gargallo" as the original author of this work.
+The file in which this license is included (LICENSE.md) must remain unmodified. Any modifications to this file will be considered a violation of this license.
+Disclaimer
+The Software is provided "as is", without warranty of any kind, express or implied, including but not limited to the warranties of merchantability, fitness for a particular purpose and noninfringement. In no event shall the authors or copyright holders be liable for any claim, damages or other liability, whether in an action of contract, tort or otherwise, arising from, out of or in connection with the Software or the use or other dealings in the Software.
+
+For further details about this license, please visit: https://huggingface.co/spaces/superdatas/free-license
